Move widget factory & styles to gui
Keep all related features in same directory.
