+++ src/gui/widget_factory.cpp
